moving classes around and refactoring low level stream classes for better comprehension
