+++ TSQL_Parser/TSQL_Parser/IO/ICharacterReader.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TSQL.IO
+{
+	interface ICharacterReader : IDisposable, IEnumerator, IEnumerable, IEnumerator<char>, IEnumerable<char>
+	{
+		
+    }
+}
