+++ TSQL_Parser/TSQL_Parser/IO/TSQLCharacterReader.IDisposable.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace TSQL
+namespace TSQL.IO
 {
 	public partial class TSQLCharacterReader : IDisposable
 	{
