--- TSQL_Parser/TSQL_Parser/TSQLCharacterReader.ICharacterReader.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TSQL
-{
-	public partial class TSQLCharacterReader
-	{
-		private interface ICharacterReader : IDisposable
-		{
-			int Read();
-        }
-	}
-}
