+++ TSQL_Parser/TSQL_Parser/TSQLTokenizer.IEnumerable.cs
@@ -37,7 +37,7 @@ namespace TSQL
 
 		void IEnumerator.Reset()
 		{
-			throw new NotSupportedException("Reset is not currently supported by the IEnumerable implementation supplied by " + GetType().FullName + ".");
+			throw new NotSupportedException("Reset is not currently supported by the IEnumerator implementation supplied by " + GetType().FullName + ".");
 		}
 
 		#endregion
