+++ TSQL_Parser/TSQL_Parser/TSQL_Parser.csproj
@@ -40,13 +40,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+		<Compile Include="IO\BufferedTextReader.cs" />
+		<Compile Include="IO\ICharacterReader.cs" />
+		<Compile Include="IO\TSQLCharacterReader.cs" />
+		<Compile Include="IO\TSQLCharacterReader.IDisposable.cs">
+			<DependentUpon>TSQLCharacterReader.cs</DependentUpon>
+		</Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Statements\Builders\ITSQLStatementBuilder.cs" />
     <Compile Include="Statements\Builders\TSQLSelectStatementBuilder.cs" />
+		<Compile Include="Statements\Builders\TSQLStatementBuilderFactory.cs" />
     <Compile Include="Statements\Builders\TSQLUnknownStatementBuilder.cs" />
     <Compile Include="Statements\TSQLSelectStatement.cs" />
     <Compile Include="Statements\TSQLStatement.cs" />
-    <Compile Include="Statements\Builders\TSQLStatementBuilderFactory.cs" />
     <Compile Include="Statements\TSQLStatementReader.cs" />
     <Compile Include="Statements\TSQLStatementReader.IDisposable.cs">
       <DependentUpon>TSQLStatementReader.cs</DependentUpon>
@@ -78,16 +84,6 @@
     </Compile>
     <Compile Include="TSQLTokenizer.IEnumerable.cs">
       <DependentUpon>TSQLTokenizer.cs</DependentUpon>
-    </Compile>
-    <Compile Include="TSQLCharacterReader.cs" />
-    <Compile Include="TSQLCharacterReader.BufferedTextReader.cs">
-      <DependentUpon>TSQLCharacterReader.cs</DependentUpon>
-    </Compile>
-    <Compile Include="TSQLCharacterReader.ICharacterReader.cs">
-      <DependentUpon>TSQLCharacterReader.cs</DependentUpon>
-    </Compile>
-    <Compile Include="TSQLCharacterReader.IDisposable.cs">
-      <DependentUpon>TSQLCharacterReader.cs</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
